perf_tests/requirements.txt
httpx==0.27.2
PyYAML==6.0.2
anyio==4.7.0
